requirements.txt
pandas
streamlit
python-dotenv
altair
requests
numpy
datetime
